api/requirements.txt
wheel
opencv-python-headless
fastapi
uvicorn
slowapi
aiofiles
python-multipart
RPi.GPIO